+++ sdk/cosmos/azure-cosmos/azure/cosmos/_location_cache.py
@@ -25,6 +25,7 @@ DatabaseAccount with multiple writable and readable locations.
 import collections
 import logging
 import time
+from urllib.parse import urlparse
 
 from . import documents
 from . import http_constants
@@ -100,7 +101,16 @@ def get_endpoints_by_location(new_locations,
                         regional_object = RegionalEndpoint(region_uri, region_uri)
                         # if it is for writes, then we update the previous to default_endpoint
                         if writes and not use_multiple_write_locations:
-                            regional_object = RegionalEndpoint(region_uri, default_regional_endpoint.get_current())
+                            # if region_uri is different than global endpoint set global endpoint
+                            # as fallback
+                            # else construct regional uri
+                            if region_uri != default_regional_endpoint.get_current():
+                                regional_object.set_previous(default_regional_endpoint.get_current())
+                            else:
+                                constructed_region_uri =  LocationCache.GetLocationalEndpoint(
+                                    default_regional_endpoint.get_current(),
+                                    new_location["name"])
+                                regional_object.set_previous(constructed_region_uri)
 
                     # pass in object with region uri , last known good, curr etc
                     endpoints_by_location.update({new_location["name"]: regional_object})
@@ -439,3 +449,29 @@ class LocationCache(object):  # pylint: disable=too-many-public-methods,too-many
                 and request.operation_type == documents._OperationType.ExecuteJavaScript
             )
         )
+
+    @staticmethod
+    def GetLocationalEndpoint(default_endpoint, location_name):
+        # For default_endpoint like 'https://contoso.documents.azure.com:443/' parse it to
+        # generate URL format. This default_endpoint should be global endpoint(and cannot
+        # be a locational endpoint) and we agreed to document that
+        endpoint_url = urlparse(default_endpoint)
+
+        # hostname attribute in endpoint_url will return 'contoso.documents.azure.com'
+        if endpoint_url.hostname is not None:
+            hostname_parts = str(endpoint_url.hostname).lower().split(".")
+            if hostname_parts is not None:
+                # global_database_account_name will return 'contoso'
+                global_database_account_name = hostname_parts[0]
+
+                # Prepare the locational_database_account_name as contoso-EastUS for location_name 'East US'
+                locational_database_account_name = global_database_account_name + "-" + location_name.replace(" ", "")
+
+                # Replace 'contoso' with 'contoso-EastUS' and return locational_endpoint
+                # as https://contoso-EastUS.documents.azure.com:443/
+                locational_endpoint = default_endpoint.lower().replace(
+                    global_database_account_name, locational_database_account_name, 1
+                )
+                return locational_endpoint
+
+        return None
